requirements.txt
fastapi>=0.115.0
httpx>=0.28.1
Jinja2>=3.1.4
lap>=0.5.13
networkx>=3.4.2
opencv-python>=4.13.0.92
Pillow>=10.4.0
PyYAML>=6.0.2
python-multipart>=0.0.9
qrcode>=7.4.2
requests>=2.32.3
ultralytics>=8.4.40
uvicorn>=0.30.6
